pyproject.toml
[project]
name = "secondbrain"
version = "0.1.0"
description = "Self-hosted Telegram bot for capturing and organizing project ideas."
readme = "README.md"
authors = [
    { name = "Esteban Soler", email = "esteban.soler@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "click>=8",
    "openai>=1.40",
    "python-telegram-bot>=21",
    "pyyaml>=6",
    "rapidfuzz>=3",
    "sqlalchemy>=2",
]

[project.scripts]
second-brain = "secondbrain.cli:main"

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.6",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]
